loan-underwriting-system/requirements.txt
fastmcp>=0.1.0
httpx>=0.25.0
asyncio
pydantic>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.11
langchain-community>=0.0.13
ollama>=0.1.7